Add react-router. Add left-hand nav component.
