+++ examples/package.json
@@ -51,7 +51,8 @@
     "react-dom": "^0.14.0",
     "react-highcharts": "^7.0.0",
     "react-redux": "^3.1.0",
-    "react-tap-event-plugin": "0.2.1",
+    "react-router": "^2.0.0",
+    "react-tap-event-plugin": "^0.2.1",
     "redux": "^3.0.2",
     "redux-thunk": "^1.0.0",
     "safe-framework": "0.0.6"
